% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.FDAimage.R
\name{data.FDAimage}
\alias{data.FDAimage}
\title{Generating simulation examples}
\usage{
data.FDAimage(n, Z, ind.inside, sigma, rho = 0.5, iter = 2019,
  lambda1 = 0.03, lambda2 = 0.006)
}
\arguments{
\item{n}{The number of images.
\cr}

\item{Z}{The cooridinates of dimension \code{npix} by two. Each row is the coordinates of a pixel/voxel.
\cr}

\item{ind.inside}{The indices of the pixels/voxels which are inside the boundary.
\cr}

\item{sigma}{The level of white noise.
\cr}

\item{rho}{The correlation used to generate the explanatory variales -- default is 0.5.
\cr}

\item{iter}{The seed used to generate the simulation examples -- default is 2019.
\cr}

\item{lambda1&lambda2}{The eigenvalues used to generate the simulation examples -- default are 0.03 and 0.006, respectively.
\cr}
}
\value{
\item{Y}{The response images.}
\item{X}{The covariates.}
\item{Z}{The coordinates of the pixels/voxels.}
\item{beta.true}{The true coefficient functions.}
\cr
}
\description{
This function is used to generate the simulation examples for image-on-scalar regression.
}
\details{
This R package is the implementation program for manuscript entitled ``Multivariate Spline Estimation and Inference for Varying Coeffiient Models with Imaging Data" by Shan Yu, Guannan Wang, Li Wang and Lijian Yang.
}
