% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.FDAimage.R
\name{plot.FDAimage}
\alias{plot.FDAimage}
\title{Produces coefficient function plots for a fitted "FDAimage" object.}
\usage{
\method{plot}{FDAimage}(mfit)
}
\arguments{
\item{mfit}{Fitted ``FDAimage" object.
\cr}
}
\value{
None
}
\description{
This function produces the images of the estimated coefficient function for image-on-scalar regression.
}
\details{
This R package is the implementation program for manuscript entitled ``Multivariate Spline Estimation and Inference for Varying Coeffiient Models with Imaging Data" by Shan Yu, Guannan Wang, Li Wang and Lijian Yang.
}
