% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.FDAimage.R
\name{predict.FDAimage}
\alias{predict.FDAimage}
\title{Make predictions from a fitted multivariate spline image-on-scalar regression model.}
\usage{
\method{predict}{FDAimage}(mfit, Xpred, Zpred)
}
\arguments{
\item{mfit}{Fitted ``FDAimage" object.
\cr}

\item{Xpred}{The design matrix for prediction.
\cr}

\item{Zpred}{The cooridinates for prediction.
\cr}
}
\value{
A matrix of predicted images is returned.
}
\description{
This function is used to make predictions of the multivariate spline image-on-scalar regression models.
}
\details{
This R package is the implementation program for manuscript entitled "Multivariate Spline Estimation and Inference for Image-on-Scalar Regression" by Shan Yu, Guannan Wang, Li Wang and Lijian Yang.
}
