% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc.FDAimage.R
\name{scc.FDAimage}
\alias{scc.FDAimage}
\title{Simultaneous confidence corridors for Image-on-Scalar Regression}
\usage{
scc.FDAimage(Y, X, Z, V, Tr, d = 5, r = 1, lambda = 10^seq(-6, 6, by
  = 0.5), alpha0 = 0.05, nboot = 100, ksi = 0.01)
}
\arguments{
\item{Y}{The response images of dimension \code{n} by \code{npix}, where \code{n} is the number of observed images and \code{npix} is the number of pixels/voxels in each image. Each row is an observation of the images.
\cr}

\item{X}{The design matrix of dimension \code{n} by \code{p}, with an intercept. Each row is an observation vector.
\cr}

\item{Z}{The cooridinates of dimension \code{npix} by two. Each row is the coordinates of a pixel/voxel.
\cr}

\item{V}{The \code{N} by two matrix of vertices of a triangulation, where \code{N} is the number of vertices. Each row is the coordinates for a vertex.
\cr}

\item{Tr}{The triangulation matrix of dimention \code{nT} by three, where \code{nT} is the number of triangles in the triangulation. Each row is the indices of vertices in \code{V}.
\cr}

\item{d}{The degree of piecewise polynomials -- default is 5, and -1 represents piecewise constant.
\cr}

\item{r}{The smoothness parameter -- default is 1, and 0 \eqn{\le} \code{r} \eqn{<} \code{d}.
\cr}

\item{lambda}{The vector of the candidates of penalty parameter -- default is grid points of 10 to the power of a sequence from -6 to 6 by 0.5.
\cr}

\item{alpha0}{The nominal level of the SCC -- default is 0.05.
\cr}

\item{nboot}{The number of bootstraps -- default is 100.}
}
\value{
A list of the lower SCC and upper SCC for the coefficient functions.
}
\description{
This function is used to construct the simultaneous confidence corridors (SCC) for the coefficient functions in image-on-scalar regression.
}
\details{
This R package is the implementation program for manuscript entitled ``Multivariate Spline Estimation and Inference for Varying Coeffiient Models with Imaging Data" by Shan Yu, Guannan Wang, Li Wang and Lijian Yang.
}
